% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add.control.points}
\alias{add.control.points}
\title{Function to add bezier control points for horizontal layout.}
\usage{
add.control.points(data = NULL, strength = 0.5)
}
\arguments{
\item{data}{A \code{data.frame} containing x and y coordinates.}

\item{strength}{The proportion to move the control point along the y-axis towards the other end of the bezier curve.}
}
\value{
A \code{vector} of rescaled coordinate values.
}
\description{
Function to add bezier control points for horizontal layout.
}
\author{
David Porubsky
}
