% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiro.R
\name{add_annotation}
\alias{add_annotation}
\title{Add annotation ranges to miropeat style plot.}
\usage{
add_annotation(
  ggplot.obj = NULL,
  annot.gr = NULL,
  shape = "arrowhead",
  fill.by = NULL,
  color.palette = NULL,
  coordinate.space = "target"
)
}
\arguments{
\item{ggplot.obj}{A \code{ggplot2} object generated using \code{\link{plotMiro}} function.}

\item{annot.gr}{A \code{\link{GRanges-class}} object with a set of ranges to be added as extra annotation.}

\item{shape}{A user defined shape ranges in 'annot.gr' are visualized, either 'arrowhead' or 'rectangle'.}

\item{fill.by}{A name of an extra field present in 'annot.gr' to be used to define color scheme.}

\item{color.palette}{User defined discrete color palette as named character vector (elements = colors, names = discrete levels).}

\item{coordinate.space}{A coordinate space ranges in 'annot.gr' are reported, either 'target' or 'query'.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function takes a \code{ggplot2} object generated using \code{\link{plotMiro}} function and adds extra annotation on top of query
or target coordinates. These ranges are specified in 'annot.gr' object and are visualized either as arrowheads or rectangles.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Make a plot
plt <- plotMiro(paf.table = paf.table)
## Load target annotation file
target.annot <- system.file("extdata", "test1_target_annot.txt", package="SVbyEye")
target.annot.df <- read.table(target.annot, header = TRUE, sep = '\t', stringsAsFactors = FALSE)
target.annot.gr <- makeGRangesFromDataFrame(target.annot.df)
## Add target annotation as arrowhead
plt <- add_annotation(ggplot.obj = plt, annot.gr = target.annot.gr, coordinate.space = 'target')
## Load query annotation file
query.annot <- system.file("extdata", "test1_query_annot.txt", package="SVbyEye")
query.annot.df <- read.table(query.annot, header = TRUE, sep = '\t', stringsAsFactors = FALSE)
query.annot.gr <- makeGRangesFromDataFrame(query.annot.df)
## Add query annotation as rectangle
add_annotation(ggplot.obj = plt, annot.gr = query.annot.gr, shape = 'rectangle', coordinate.space = 'query')
## Lift target annotation to query and plot
lifted.annot.gr <- liftRangesToAlignment(gr = target.annot.gr, paf.file = paf.file, direction = 'target2query')
add_annotation(ggplot.obj = plt, annot.gr = lifted.annot.gr, shape = 'rectangle', coordinate.space = 'query')
## Add segmental duplication annotation
plt <- plotMiro(paf.table = paf.table)
sd.annot <- system.file("extdata", "test1.sd.annot.RData", package="SVbyEye")
sd.annot.gr <- get(load(sd.annot))
## Create a custom discrete levels
sd.categ <- findInterval(sd.annot.gr$fracMatch, vec = c(0.95, 0.98, 0.99))
sd.categ <- dplyr::recode(sd.categ, '0' = '<95\%', '1' = '95-98\%', '2' = '98-99\%', '3'='>=99\%')
sd.categ <- factor(sd.categ, levels=c('<95\%', '95-98\%', '98-99\%', '>=99\%'))
sd.annot.gr$sd.categ <- sd.categ
color.palette <- c('<95\%' = 'gray72', '95-98\%' = 'gray47', '98-99\%' = '#cccc00', '>=99\%' = '#ff6700')
## Add annotation to the plot
add_annotation(ggplot.obj = plt, annot.gr = sd.annot.gr, fill.by = 'sd.categ', color.palette = color.palette, coordinate.space = 'target')

}
\author{
David Porubsky
}
