% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssemblyAlignments.R
\name{bed2ranges}
\alias{bed2ranges}
\title{Load an de novo assembly aligned to the reference in BED file into a \code{\link{GRanges-class}} object.}
\usage{
bed2ranges(
  bedfile = NULL,
  index = NULL,
  min.mapq = 10,
  min.align = 10000,
  min.ctg.size = 5e+05,
  max.gap = 1e+05,
  report.ctg.ends = FALSE
)
}
\arguments{
\item{bedfile}{A BED file of contig alignments to a reference genome.}

\item{index}{A unique identifier to be added as an 'ID' field.}

\item{min.mapq}{A minimum mapping quality of alignments reported in submitted BED file.}

\item{min.align}{A minimum length of an aligned sequence to a reference genome.}

\item{min.ctg.size}{A minimum length a final contig after gaps are collapsed.}

\item{max.gap}{A maximum length of a gap within a single contig alignments to be collapsed.}

\item{report.ctg.ends}{Set to \code{TRUE} if alignment ends of each contig to the reference should reported as well.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will take a BED file of contig alignments to a reference genome and converts them
into a \code{\link{GRanges-class}} object. This function also aims to collapse single unique contigs
that are aligned in multiple pieces after the alignment to the reference.
}
\author{
David Porubsky
}
