% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakPafAlignment.R
\name{breakPafAlignment}
\alias{breakPafAlignment}
\title{Function to break PAF alignment into matching bases between query and target sequence.
In addition, locations of inserted bases in query and target sequence can be reported as well.}
\usage{
breakPafAlignment(
  paf.aln = NULL,
  min.deletion.size = 50,
  min.insertion.size = 50,
  collapse.mismatches = TRUE,
  report.sv = TRUE
)
}
\arguments{
\item{paf.aln}{A \code{data.frame} or \code{tibble} containing a single PAF record with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}

\item{collapse.mismatches}{Set to \code{TRUE} if mismatches should be collapsed in order expand matched regions.}

\item{report.sv}{Set to \code{TRUE} if to report also ranges of deleted and inserted bases.}
}
\value{
A \code{list} of \code{tibble} objects storing matched ('M') alignments as well as structurally variable ('SV') bases if 'report.sv' is TRUE.
}
\description{
Function to break PAF alignment into matching bases between query and target sequence.
In addition, locations of inserted bases in query and target sequence can be reported as well.
}
\examples{
## Get PAF to break
paf.file <- system.file("extdata", "test2.paf", package = "SVbyEye")
## Read in PAF alignment
paf.aln <- readPaf(paf.file = paf.file)
## Break PAF alignment at indels of 1 kbp and longer
breakPafAlignment(paf.aln = paf.aln, min.deletion.size = 1000, min.insertion.size = 1000)

}
\author{
David Porubsky
}
