% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarUtils.R
\name{cigar2ranges}
\alias{cigar2ranges}
\title{Function to load PAF alignments into a set of genomic ranges.}
\usage{
cigar2ranges(
  paf.file = NULL,
  coordinate.space = "reference",
  min.insertion.size = 50,
  min.deletion.size = 50,
  collapse.mismatches = TRUE
)
}
\arguments{
\item{paf.file}{A character string containing alignment represented as a CIGAR string.}

\item{coordinate.space}{A used defined coordinate space given CIGAR should be parsed against, either 'reference' or 'query'.}

\item{min.insertion.size}{A minimum size (in basepairs) of an insertion to be retained.}

\item{min.deletion.size}{A minimum size (in basepairs) of a deletion to be retained.}

\item{collapse.mismatches}{Set to \code{TRUE} if mismatches should be collapsed in order expand matched regions.}
}
\value{
A \code{\link{GRanges}} object.
}
\description{
Function to load PAF alignments into a set of genomic ranges.
}
\author{
David Porubsky
}
