% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collapsePaf}
\alias{collapsePaf}
\title{Collapse PAF query and target ranges based withing a group}
\usage{
collapsePaf(paf.table, collapse.by = NULL)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{collapse.by}{A user defined column name present in `paf.table` to serve as grouping variable.}
}
\value{
A \code{tibble} of collapsed PAF alignments.
}
\description{
This function takes PAF alignments stored in using \code{tibble} object and collapse them
based on grouping variable defined in `collapse.by` that have to be a valid column name
present in `paf.table`.
}
\examples{
## Get PAF to plot ##
paf.file <- system.file("extdata", "test1.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Split PAF alignments into user defined bins
paf.table <- pafToBins(paf.table = paf.table, binsize = 1000)
## Collapse PAF alignments by bin id
collapsePaf(paf.table = paf.table, collapse.by = "bin.id")

}
\author{
David Porubsky
}
