% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakPafAlignment.R
\name{expandPaf}
\alias{expandPaf}
\title{A wrapper function for \code{\link{breakPafAlignment}} expanding multiple PAF alignments
into a set of matching bases between query and target sequence.}
\usage{
expandPaf(
  paf.table = NULL,
  min.deletion.size = 50,
  min.insertion.size = 50,
  collapse.mismatches = TRUE,
  report.sv = TRUE
)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single PAF record with 12 mandatory columns.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}

\item{collapse.mismatches}{Set to \code{TRUE} if mismatches should be collapsed in order expand matched regions.}

\item{report.sv}{Set to \code{TRUE} if to report also ranges of deleted and inserted bases.}
}
\value{
A \code{list} of \code{tibble} objects storing matched ('M') alignments as well as structurally vairable ('SV') bases if 'report.sv' is TRUE.
}
\description{
A wrapper function for \code{\link{breakPafAlignment}} expanding multiple PAF alignments
into a set of matching bases between query and target sequence.
}
\author{
David Porubsky
}
