% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssemblyAlignments.R
\name{fillGaps}
\alias{fillGaps}
\title{This function will takes in a \code{\link{GRanges-class}} object of alignments of a single contig to a reference
and collapses gaps in alignment based user specified maximum allowed gap.}
\usage{
fillGaps(gr, max.gap = 1e+05)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of regions of a single contigs aligned to a reference.}

\item{max.gap}{A maximum length of a gap within a single contig alignments to be collapsed.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will takes in a \code{\link{GRanges-class}} object of alignments of a single contig to a reference
and collapses gaps in alignment based user specified maximum allowed gap.
}
\author{
David Porubsky
}
