% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPaf.R
\name{filterPaf}
\alias{filterPaf}
\title{Filter PAF alignments.}
\usage{
filterPaf(
  paf.table,
  min.mapq = 10,
  min.align.len = 1000,
  min.align.n = 1,
  is.selfaln = FALSE,
  min.selfaln.dist = 0
)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{min.mapq}{Minimum mapping quality to retain.}

\item{min.align.len}{Minimum alignment length to retain.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{is.selfaln}{Set to \code{TRUE} if processing alignments of a given sequence to itself.}

\item{min.selfaln.dist}{Keep alignment pairs with this or larger distance from each other (Applied only to FASTA self-alignments).}
}
\value{
A \code{tibble} of filtered PAF alignments.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and perform
user defined filtering of input alignments based on mapping quality, alignment length, and
minimum alignments between target and query.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Filter PAF alignments based on desired target region
filterPaf(paf.table = paf.table, min.align.len = 50000)

}
\author{
David Porubsky
}
