% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiro.R
\name{flipQueryAnnotation}
\alias{flipQueryAnnotation}
\title{Flip query annotation ranges}
\usage{
flipQueryAnnotation(paf.table, query.annot.gr = NULL)
}
\arguments{
\item{query.annot.gr}{A \code{\link{GRanges-class}} object with a set of ranges in query coordinates. See function
\code{\link{liftRangesToAlignment}} if coordinates need to be lifted from target space.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and postprocessed using
\code{\link{flipPaf}} function. In case the PAF alignments were flipped ranges defined in 'query.annot.gr'
will be flipped accordingly to match query coordinates defined in 'paf.table'.
}
\examples{
## Get PAF to process
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Flip PAF alignments
paf.table <- flipPaf(paf.table = paf.table, force=TRUE)
## Load query annotation file
query.annot <- system.file("extdata", "test1_query_annot.txt", package="SVbyEye")
query.annot.df <- read.table(query.annot, header = TRUE, sep = '\t', stringsAsFactors = FALSE)
query.annot.gr <- makeGRangesFromDataFrame(query.annot.df)
## Synchronize orientation of query annotation file with flipped PAF alignments
flipQueryAnnotation(paf.table = paf.table, query.annot.gr=query.annot.gr)

}
\author{
David Porubsky
}
