% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arrowhead.R
\name{geom_arrowhead}
\alias{geom_arrowhead}
\title{A 'ggplot2' geom to draw genomic ranges as arrowheads.}
\usage{
geom_arrowhead(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  arrowhead_height = grid::unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, }{etc... As standard for ggplot2.}

\item{arrowhead_height}{A \code{grid::unit()} object providing the height of the arrowhead.  \link{Default: 3 mm}.}
}
\description{
\code{geom_arrowhead()} draws ranges defined by \code{xmin} and \code{xmax} as triangular polygon.
draws genomic ranges as arrowheads, allowing to draw for instance segmental
duplication maps.
}
\details{
This geom draws triangular polygons as arrowheads between defined start and end coordinates.
Intended application of this geom is to visualize genomic coordinates defined by start and end
position.
}
\section{Aesthetics}{

\code{geom_roundrect()} require or can take the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{xmin}
\item \strong{xmax}
\item \strong{y}
\item color
\item linewidth
\item linetype
\item alpha
\item fill
\item size
}
}

\examples{
## Create example data.frame to plot
plt.df <- data.frame(xmin=c(10, 100, 200),
                    xmax=c(100, 190, 400)
                    )
## Plot rectangles with rounded edges
ggplot2::ggplot(plt.df) + 
 geom_arrowhead(ggplot2::aes(xmin=xmin, xmax=xmax, y=1))

}
\author{
David Porubsky
}
