% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_roundrect.R
\name{geom_roundrect}
\alias{geom_roundrect}
\title{A 'ggplot2' geom to draw genomic annotations as round rectangles}
\usage{
geom_roundrect(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  rect_height = grid::unit(3, "mm"),
  radius = grid::unit(1, "mm"),
  ...
)
}
\description{
\code{geom_roundrect()} draws genomic ranges as round rectangles, allowing to draw for instance ...
}
\details{
This geom draws ...
}
