% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wide_arc.R
\name{geom_wide_arc}
\alias{geom_wide_arc}
\title{A 'ggplot2' geom to draw arcs between genomic alignments.}
\usage{
geom_wide_arc(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  stat = "wide_arc",
  position = "identity",
  n = 100,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, }{etc... As standard for ggplot2.}
}
\description{
\code{geom_wide_arc()} draws wide polygons between two sets of start and end coordinates.
}
\details{
This geom is intended to draws wide arc polygons between self-alignments defined in PAF format.
Such alignments can be directly visualized using a wrapper function \code{\link{plotSelf}}.
Input data is a \code{data.frame} object that contains required aesthetics \code{x} and \code{y} coordinates as well as
\code{group} field that is required in order to determine which coordinates represent a single alignment.
}
\section{Aesthetics}{

\code{geom_wide_arc()} require or can take the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{group}
\item color
\item linewidth
\item linetype
\item alpha
\item fill
\item size
}
}

\examples{
## Create example data.frame to plot
plt.df <- data.frame(x=c(100, 500, 200, 1000),
                    y=0,
                    group=1)
## Make a plot 
ggplot2::ggplot(plt.df) + 
 geom_wide_arc(ggplot2::aes(x=x, y=y, group=group), alpha=0.5)

}
\seealso{
\code{\link[=plotSelf]{plotSelf()}}
}
\author{
David Porubsky
}
