% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequenceContent.R
\name{getNucleotideContent}
\alias{getNucleotideContent}
\title{Get sequence content from one or multiple FASTA sequences.}
\usage{
getNucleotideContent(
  fasta.seq,
  sequence.pattern = NULL,
  nucleotide.content = NULL
)
}
\arguments{
\item{fasta.seq}{A \code{\link{DNAStringSet-class}} object containing one or multiple FASTA sequences.}

\item{sequence.pattern}{A user defined DNA sequence pattern which occurrences will be counted in submitted `fasta.seq`.
(e.g. set `sequence.pattern` to 'GA' to obtain counts of all 'GA' occurrences per FASTA sequence).}

\item{nucleotide.content}{A user defined nucleotides which total content will be counted in submitted `fasta.seq`.
(e.g. to obtain 'GC' content set `nucleotide.content` to 'GC')}
}
\value{
A \code{tibble} of required nucleotide/sequence content
}
\description{
This function takes a single or set of FASTA sequences and for each sequence will report counts and frequencies of user defined
`sequence.pattern` (such as exact DNA pattern, e.g. 'GA') or `nucleotide.content` (such as sequence GC content).
}
\examples{
## Get PAF to process ##
paf.file <- system.file("extdata", "test_getFASTA.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Get FASTA to process
asm.fasta <- system.file("extdata", "test_getFASTA_query.fasta", package = "SVbyEye")
## Read in FASTA sequence
fasta.seq <- paf2FASTA(paf.table = paf.table, asm.fasta = asm.fasta)
## Report sequence and nucleotide content for a given FASTA sequence
getNucleotideContent(fasta.seq = fasta.seq, sequence.pattern = "AT", nucleotide.content = "GC")

}
\author{
David Porubsky
}
