% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftRangesToAlignment_deprecated.R
\name{liftRangesToAlignment_deprecated}
\alias{liftRangesToAlignment_deprecated}
\title{Function to lift local sequence coordinates to the alignment of a given sequence to the reference.}
\usage{
liftRangesToAlignment_deprecated(gr = NULL, paf.file = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object containing single or multiple ranges in local sequence coordinates.}

\item{paf.file}{A path to a PAF file containing alignments of a local sequence to the reference sequence.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Function to lift local sequence coordinates to the alignment of a given sequence to the reference.
}
\author{
David Porubsky
}
