% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBins.R
\name{makeBins}
\alias{makeBins}
\title{Split genome into user defined bins.}
\usage{
makeBins(
  bsgenome = NULL,
  fai = NULL,
  chromosomes = NULL,
  binsize = 1e+05,
  stepsize = binsize/2
)
}
\arguments{
\item{bsgenome}{A reference genome to get lengths of genomic sequences (eg. GRCh38).}

\item{fai}{A FASTA index to get lengths of genomic sequences.}

\item{chromosomes}{A user defined set of chromosomes for binning (eg. 'chr1')}

\item{binsize}{A size of the genomic bin to split genome into.}

\item{stepsize}{A size of the genomic interval to move each bin. For non-overlapping bins use the same size as binsize.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
This function splid genome into a equaly-sized user defined genomic intervals.
}
\author{
David Porubsky
}
