% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskFasta.R
\name{maskFASTA}
\alias{maskFASTA}
\title{Mask FASTA sequence at defined regions.}
\usage{
maskFASTA(
  fasta.file,
  mask.ranges = NULL,
  invert = FALSE,
  mask.character = "N",
  fasta.save = NULL
)
}
\arguments{
\item{fasta.file}{A path to a FASTA file to be masked.}

\item{mask.ranges}{A \code{\link{IRanges-class}} object of coordinates to be masked in input FASTA.}

\item{invert}{If set \code{TRUE} ranges defined in 'mask.ranges' will be kept while the rest of the FASTA will be masked.}

\item{mask.character}{A single character to be used for masking (Default : 'N').}

\item{fasta.save}{A path to a filename where to store final FASTA file.}
}
\description{
Mask FASTA sequence at defined regions.
}
\examples{
## FASTA sequence to mask
fasta.file <- system.file("extdata", "test4_query.fasta", package = "SVbyEye")
## Define ranges to mask
mask.ranges <- IRanges::IRanges(start = c(11, 1981), end = c(20, 1990))
## Mask FASTA sequence at defined ranges
maskFASTA(fasta.file = fasta.file, mask.ranges = mask.ranges)
## Mask FASTA sequence except of defined ranges
maskFASTA(fasta.file = fasta.file, mask.ranges = mask.ranges, invert = TRUE)

}
\author{
David Porubsky
}
