% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mirrorRanges}
\alias{mirrorRanges}
\title{Mirror/reflect genomic ranges given the sequence length.}
\usage{
mirrorRanges(gr, seqlength = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with one or more ranges to be mirrored/reflected given the sequence length.}

\item{seqlength}{A \code{numeric} value containing the sequence length from where the submitted ranges originate from.}
}
\value{
A \code{\link{GRanges-class}} object with mirrored coordinates.
}
\description{
This function flips set of ranges in a mirrored fashion.
}
\author{
David Porubsky
}
