% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucmerToPaf.R
\name{nucmer2PAF}
\alias{nucmer2PAF}
\title{Read nucmer coordinates file in to PAF formatted table}
\usage{
nucmer2PAF(nucmer.coords = NULL)
}
\arguments{
\item{nucmer.coords}{A path to a nucmer coordinates file containing alignments to be loaded.}
}
\description{
This function takes alignments produced by nucmer (parameters: --mum --coords) and loads
reported coordinates file (suffix .coords) into a PAF formatted table (see PAF specification).
}
\examples{
## Get nucmer coordinates to read in ##
nucmer.coords <- system.file("extdata", "HG00438_1_nucmer.coords", package = "SVbyEye")
## Read in nucmer coordinates
nucmer2PAF(nucmer.coords=nucmer.coords)

}
\author{
David Porubsky
}
