% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paf2continuousScale}
\alias{paf2continuousScale}
\title{Define shift of genomic coordinates in order to get continuous coordinate scale.}
\usage{
paf2continuousScale(paf.table)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}
}
\value{
A \code{tibble} of PAF alignments reported with a continuous coordinates.
}
\description{
This function takes a PAF alignments loaded by \code{\link{readPaf}} function and defines how much to shift genomic coordinates
in order to get continuous coordinates in cases when there more than one query or target sequences in the input PAF alignments.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test_ava.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Filter PAF
paf.table <- paf.table[paf.table$t.name == "HG03453_2", ]
## Define shift in genomic coordiantes to get continuous scale
paf2continuousScale(paf.table)

}
\author{
David Porubsky
}
