% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{paf2coords}
\alias{paf2coords}
\title{Read and filter PAF input file}
\usage{
paf2coords(
  paf.file,
  min.mapq = 10,
  min.align.len = 1000,
  min.align.n = 10,
  seqname.grep = NULL
)
}
\arguments{
\item{paf.file}{A \code{data.frame} containing x and y coordinates.}

\item{min.mapq}{Minimum mapping quality to retain for PAF alignment file.}

\item{min.align.len}{Minimum alignment length to retain for PAF alignment file.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{seqname.grep}{Retain only a specific sequence/region name with a given character string.}
}
\value{
A \code{vector} of rescaled coordinate values.
}
\description{
This function takes PAF output file from minimap2 alignemts, loads the file and
perform user defined filtering of input alignements based on mapping quality and
alignment length.
}
\author{
David Porubsky
}
