% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafToGaps.R
\name{paf2gaps}
\alias{paf2gaps}
\title{Report gaps between PAF alignments.}
\usage{
paf2gaps(paf.table, min.gap.diff = NULL)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{min.gap.diff}{A user defined minimum gap size difference in target and query coordinates (and vice versa)
to be retained.}
}
\value{
A \code{tibble} of PAF alignment gaps.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and then reports all gaps
in subsequent target and query alignments and classify them as either deletions ('D') or insertions ('I').
}
\examples{
## Get PAF to process ##
paf.file <- system.file("extdata", "test1.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Report gaps between PAF alignments
paf2gaps(paf.table = paf.table)

}
\author{
David Porubsky
}
