% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{paf2ranges}
\alias{paf2ranges}
\title{Load a de novo assembly aligned to the reference in PAF format into a \code{\link{GRanges-class}} object.}
\usage{
paf2ranges(
  paf.file = NULL,
  index = NULL,
  min.mapq = 10,
  min.aln.width = 10000,
  min.ctg.size = 5e+05,
  report.ctg.ends = FALSE
)
}
\arguments{
\item{paf.file}{A BED file of contig alignments to a reference genome.}

\item{min.ctg.size}{A minimum length a final contig after gaps are collapsed.}

\item{report.ctg.ends}{Set to \code{TRUE} if aligned position of each contig ends should be reported}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will take a PAF file of contig alignments to a reference genome and converts them
into a \code{\link{GRanges-class}} object.
}
\author{
David Porubsky
}
