% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiro_wrapper.R, R/wrappers.R
\name{plotMiro}
\alias{plotMiro}
\title{Visualize PAF alignments.}
\usage{
plotMiro(
  paf.file = paf.file,
  min.mapq = 10,
  min.align.len = 100,
  min.align.n = 1,
  min.deletion.size = NULL,
  min.insertion.size = NULL,
  sd.annot = NULL,
  drop.self.align = FALSE
)

plotMiro(
  paf.file = paf.file,
  min.mapq = 10,
  min.align.len = 100,
  min.align.n = 1,
  min.deletion.size = NULL,
  min.insertion.size = NULL,
  sd.annot = NULL,
  drop.self.align = FALSE
)
}
\arguments{
\item{paf.file}{A path to a PAF file containing alignments to be loaded.}

\item{min.mapq}{Minimum mapping quality to retain.}

\item{min.align.len}{Minimum alignment length to retain.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}

\item{sd.annot}{A \code{\link[GenomicRanges]{GRanges}} object containing segmental duplication coordinates.}

\item{drop.self.align}{Remove alignments of a given sequence to itself.}

\item{color.by}{}

\item{flip.alignment}{Set to \code{TRUE} if query PAF alignments should be flipped.}
}
\value{
A \code{ggplot2} object or \code{list} of  \code{ggplot2} objects.

A \code{list} of miropeat style plots.
}
\description{
This function takes PAF output file from minimap2 alignments, and visualize the alignments
in miropeat style.

This function takes PAF output file from minimap2 alignments, and visualize the alignments
in miropeat style.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Make a plot
## Color by alignment directionality
plotMiro(paf.file = paf.file, color.by = 'direction')
## Color by fraction of matched bases in each alignment
plotMiro(paf.file = paf.file, color.by = 'fraction.matches')

}
\author{
David Porubsky
}
