% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{plotMiro_deprecated}
\alias{plotMiro_deprecated}
\title{Plot PAF file as miropeat alignments}
\usage{
plotMiro_deprecated(
  paf.file = paf.file,
  min.mapq = 10,
  min.align.len = 100,
  min.align.n = 1,
  min.deletion.size = NULL,
  min.insertion.size = NULL,
  sd.annot = NULL,
  drop.self.align = FALSE
)
}
\arguments{
\item{sd.annot}{A \code{\link[GenomicRanges]{GRanges}} object containing segmental duplication coordinates.}
}
\value{
A \code{list} of miropeat style plots.
}
\description{
This function takes PAF output file from minimap2 alignments, and visualize the alignments
in miropeat style.
}
\author{
David Porubsky
}
