% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariantDensity.R
\name{plotVariantDensity}
\alias{plotVariantDensity}
\title{Plot genome-wide density of single nucleotide or larger variants across the genome or separate chromosomes.}
\usage{
plotVariantDensity(
  infile = NULL,
  chromosomes = NULL,
  binsize = 20000,
  stepsize = 2e+05,
  layout = "horizontal",
  reverse.x = FALSE,
  title = NULL,
  blacklist = NULL,
  bsgenome = NULL,
  ref.fasta = NULL,
  gap.cent.bed = NULL
)
}
\arguments{
\item{infile}{A genome positions of variants to be visualized as binned density, thus far only BED file format is allowed.}

\item{chromosomes}{A user defined set of chromosomes for binning (eg. 'chr1')}

\item{binsize}{A size of the genomic bin to split genome into.}

\item{stepsize}{A size of the genomic interval to move each bin. For non-overlapping bins use the same size as binsize.}

\item{layout}{An overall layout of the ideogram, either 'horizontal' or 'vertical'.}

\item{reverse.x}{Set to \code{TRUE} if x-axis should be horizontally reversed.}

\item{title}{A character string to be used as plot title.}

\item{blacklist}{A \code{\link{GRanges-class}} object of regions to be removed.}

\item{bsgenome}{A \code{\link{BSgenome-class}} object to provide chromosome lengths for plotting.}

\item{ref.fasta}{A user defined reference FASTA file from which chromosome lengths will be extracted.}

\item{gap.cent.bed}{A set of ranges in BED format to be highlighted as centromeres and/or gaps as white rectangles.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genome-wide density of single nucleotide or larger variants across the genome or separate chromosomes.
}
\author{
David Porubsky
}
