% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPaf.R
\name{processPafTags}
\alias{processPafTags}
\title{Process PAF specific alignment tags.}
\usage{
processPafTags(paf.tags, restrict.paf.tags = c("NM", "cg"))
}
\arguments{
\item{paf.tags}{...}

\item{restrict.paf.tags}{Define a set of PAF tag ids (e.g. NM, cg) to be reported in the output.}
}
\description{
Process PAF specific alignment tags.
}
\author{
David Porubsky
}
