% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGaf.R
\name{readGaf}
\alias{readGaf}
\title{Read GAF from an input file}
\usage{
readGaf(gaf.file = NULL)
}
\arguments{
\item{gaf.file}{A path to a GAF file containing graph alignments.}
}
\description{
This function takes an GAF output file from minigraph and loads the file and
parse all the alignments.
}
\author{
David Porubsky
}
