% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPaf.R
\name{readPaf}
\alias{readPaf}
\title{Read PAF from an input file}
\usage{
readPaf(
  paf.file = NULL,
  include.paf.tags = TRUE,
  restrict.paf.tags = c("NM", "cg")
)
}
\arguments{
\item{paf.file}{A path to a PAF file containing alignments to be loaded.}

\item{include.paf.tags}{Set to \code{TRUE} if all additional PAF alignment tags should be included in the output.}

\item{restrict.paf.tags}{Define a set of PAF tag ids (e.g. NM, cg) to be reported in the output.}
}
\description{
This function takes an PAF output file from minimap2 and  loads the file along
with user defined set of additional alignment tags (see PAF specification).
}
\author{
David Porubsky
}
