% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSelfAlignment.R
\name{readSelfAlignments}
\alias{readSelfAlignments}
\title{Read self-alignments as \code{\link{GRanges-class}} object.}
\usage{
readSelfAlignments(
  aln.coords = NULL,
  format = "nucmer",
  min.align.len = 1000,
  min.align.dist = 1000,
  collapse.overlaps = TRUE,
  break.paf.aln = TRUE,
  min.deletion.size = 1000,
  min.insertion.size = 1000
)
}
\arguments{
\item{aln.coords}{A path to a file containing self-alignment coordinates.}

\item{format}{Define format of the input alignment coordinates as either 'nucmer' or minimap2 'mm2', \link{default: 'nucmer'}}

\item{min.align.len}{}

\item{min.align.dist}{Keep alignment pairs with this or larger distance from each other.}

\item{collapse.overlaps}{Set to \code{TRUE} to merge overlapping pair of alignments with the same relative orientation.}

\item{break.paf.aln}{Set to \code{TRUE} in order to split CIGAR string at insertions and deletions.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}
}
\value{
A \code{list} of \code{\link{GRanges-class}} objects.
}
\description{
This function reads FASTA self-alignments generated by 'nucmer' or 'minimap2' aligner.
This function considers proximal (leftmost) duplication as a query(s1)
and corresponding distal (rightmost) duplication as a target (s2)
}
\author{
David Porubsky
}
