% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssemblyAlignments.R
\name{reportContigCoverage}
\alias{reportContigCoverage}
\title{This function will take in a \code{\link{GRanges-class}} or \code{\link{GRangesList-class}} object of genomic ranges of a
single a contig aligned to a reference and reports coverage of regions that overlaps each other.}
\usage{
reportContigCoverage(ranges, id.col = NULL)
}
\arguments{
\item{ranges}{A \code{\link{GRanges-class}} or \code{\link{GRangesList-class}} object of regions of a single or multiple contigs aligned to a reference.}

\item{id.col}{A column number from the original \code{\link{GRanges-class}} object to be reported as an unique ID.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will take in a \code{\link{GRanges-class}} or \code{\link{GRangesList-class}} object of genomic ranges of a
single a contig aligned to a reference and reports coverage of regions that overlaps each other.
}
\author{
David Porubsky
}
