% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssemblyAlignments.R
\name{reportGaps}
\alias{reportGaps}
\title{This function will takes in a \code{\link{GRanges-class}} object of alignments of a single contig to a reference
and reports all alignment gaps.}
\usage{
reportGaps(gr, id.col = 1)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of regions of a single contigs aligned to a reference.}

\item{id.col}{A column number from the original \code{\link{GRanges-class}} object to be reported as na unique ID.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will takes in a \code{\link{GRanges-class}} object of alignments of a single contig to a reference
and reports all alignment gaps.
}
\author{
David Porubsky
}
