% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssemblyAlignments.R
\name{reportOverlapBases}
\alias{reportOverlapBases}
\title{This function will take in a \code{\link{GRanges-class}} object of genomic ranges and enumerate the number of overlapping bases
with other set of user defined 'query' genomic ranges.}
\usage{
reportOverlapBases(ranges = NULL, query.ranges = NULL, index = NULL)
}
\arguments{
\item{ranges}{A \code{\link{GRanges-class}} object of genomic regions to get the number of overlapping bases with 'query.ranges'.}

\item{query.ranges}{A \code{\link{GRanges-class}} object of genomic regions for which one want to report overlaps.}

\item{index}{A user defined name of a column where the number of overlapping bases between 'ranges' and 'query.ranges' will be reported.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will take in a \code{\link{GRanges-class}} object of genomic ranges and enumerate the number of overlapping bases
with other set of user defined 'query' genomic ranges.
}
\author{
David Porubsky
}
