% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplots.R
\name{selfdotplot}
\alias{selfdotplot}
\title{Make a horizontal self-alignment dot plot.}
\usage{
selfdotplot(
  selfaln.gr = NULL,
  shape = "segment",
  highlight.pos = NULL,
  highlight.region = NULL,
  title = NULL
)
}
\arguments{
\item{shape}{A shape used to plot aligned sequences: Either 'segment' or 'arc'.}

\item{highlight.pos}{A single or a set of positions to be highlighted as vertical solid lines.}

\item{highlight.region}{A pair of positions to be highlighted as vertical range.}

\item{title}{A title to be added to the final dotplot.}

\item{aln.coords}{A path to a file containing self-alignment coordinates.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes self-alignment coordinates from 'nucmer' or 'minimap2' aligner and
construct a horizontal dot plot.
}
\author{
David Porubsky
}
