% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplots.R
\name{simpledotplot}
\alias{simpledotplot}
\title{Plot simple dotplot of two sequences.}
\usage{
simpledotplot(
  aln.coords = NULL,
  format = "nucmer",
  shape = "segment",
  min.align.len = 1000,
  keep.longest.aln = FALSE,
  highlight.pos = NULL,
  highlight.region = NULL,
  title = NULL,
  genome.coord = FALSE
)
}
\arguments{
\item{aln.coords}{A path to a file containing self-alignment coordinates.}

\item{shape}{A shape used to plot aligned sequences: Either 'segm' or 'point'.}

\item{keep.longest.aln}{Set to \code{TRUE} if a target sequence with a single longest alignment should be kept.}

\item{highlight.pos}{A single or a set of positions to be highlighted as vertical solid lines.}

\item{highlight.region}{A pair of positions to be highlighted as vertical range.}

\item{title}{A title to be added to the final dotplot.}

\item{genome.coords}{Set to \code{TRUE} if the target sequence should be reported in genomic coordinates.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes alignment coordinates from 'nucmer' or 'minimap2' aligner and
construct a simple dot plot.
}
\author{
David Porubsky
}
