% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wide_arc.R
\docType{data}
\name{StatWideArc}
\alias{StatWideArc}
\title{A ggproto class definition to extend the functionality of ggplot2.}
\value{
Plotting coordinates
}
\description{
A ggproto class definition to extend the functionality of ggplot2.
}
\keyword{datasets}
