% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPafAlignments.R
\name{subsetPaf}
\alias{subsetPaf}
\title{Subset PAF alignments at desired genomic range.}
\usage{
subsetPaf(paf.table, target.region = NULL)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{target.region}{A user defined target region either as character string ('chr:start-end') or as
a \code{\link{GRanges-class}} object containing a single genomic region to which PAF alignments will be
narrowed down.}
}
\value{
A \code{tibble} of subsetted PAF alignments.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and then subsets
as well as cuts PAF alignments at desired target coordinates.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test_ava.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Subset multiple PAF alignments based on desired target region
target.region <- 'HG01358_2:100000-200000'
subsetPaf(paf.table, target.region = target.region)

}
\author{
David Porubsky
}
