% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncRangesDir.R
\name{syncRangesDir}
\alias{syncRangesDir}
\title{Synchronize orientation of genomic ranges given the desired majority direction.}
\usage{
syncRangesDir(ranges, majority.strand = "+", strand.only = FALSE)
}
\arguments{
\item{ranges}{A \code{\link{GRanges}} object.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{strand.only}{If set to \code{TRUE} simple character string of strand directionality ('+', '-') is reported.}
}
\description{
This function takes a set of ranges in \code{\link{GRanges-class}} object and return the same ranges
as \code{\link{GRanges-class}} object such that total length of plus (direct, '+') and minus (minus, '-') .
oriented ranges meet the user defined `majority.strand` orientation.
}
\author{
David Porubsky
}
