% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiro_wrapper.R
\name{add_annotation}
\alias{add_annotation}
\title{Add annotation ranges to miropeat style plot.}
\usage{
add_annotation(
  ggplot.obj = NULL,
  annot.gr = NULL,
  shape = "arrowhead",
  fill.by = NULL,
  coordinate.space = "target"
)
}
\arguments{
\item{ggplot.obj}{A \code{ggplot2} object generated using \code{\link{plotMiro}} function.}

\item{annot.gr}{A \code{\link{GRanges-class}} object with a set of ranges to be added as extra annotation.}

\item{shape}{A user defined shape ranges in 'annot.gr' are visualized, either 'arrowhead' or 'rectangle'.}

\item{fill.by}{A name of an extra field present in 'annot.gr' to be used to define color scheme.}

\item{coordinate.space}{A coordinate space ranges in 'annot.gr' are reported, either 'target' or 'query'.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function takes a \code{ggplot2} object generated using \code{\link{plotMiro}} function and adds extra annotation on top of query
or target coordinates. These ranges are specified in 'annot.gr' object and are visualized either as arrowheads or rectangles.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Make a plot
plt <- plotMiro(paf.table = paf.table)
## Load target annotation file
target.annot <- system.file("extdata", "test1_target_annot.txt", package="SVbyEye")
target.annot.df <- read.table(target.annot, header = TRUE, sep = '\t', stringsAsFactors = FALSE)
target.annot.gr <- makeGRangesFromDataFrame(target.annot.df)
## Add target annotation as arrowhead
plt <- add_annotation(ggplot.obj = plt, annot.gr = target.annot.gr, coordinate.space='target')
## Load query annotation file
query.annot <- system.file("extdata", "test1_query_annot.txt", package="SVbyEye")
query.annot.df <- read.table(query.annot, header = TRUE, sep = '\t', stringsAsFactors = FALSE)
query.annot.gr <- makeGRangesFromDataFrame(query.annot.df)
## Add query annotation as rectangle
add_annotation(ggplot.obj = plt, annot.gr = query.annot.gr, shape='rectangle', coordinate.space='query')
## Lift target annotation to query and plot
lifted.annot.gr <- liftRangesToAlignment(gr = target.annot.gr, paf.file = paf.file, direction = 'target2query')
add_annotation(ggplot.obj = plt, annot.gr = lifted.annot.gr, shape='rectangle', coordinate.space='query')

}
\author{
David Porubsky
}
