% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{bed2ranges}
\alias{bed2ranges}
\title{Load a de novo assembly aligned to the reference in BED format into a \code{\link{GRanges-class}} object.}
\usage{
bed2ranges(
  bed.file = NULL,
  index = NULL,
  min.mapq = 10,
  min.aln.width = 10000,
  max.gap = 1e+05,
  report.mapped.ends = FALSE
)
}
\arguments{
\item{bed.file}{A BED file of contig alignments to a reference genome.}

\item{index}{A unique identifier to be added as an 'ID' field.}

\item{min.mapq}{A minimum mapping quality of alignments reported in submitted BED file.}

\item{min.aln.width}{A minimum width of an aligned sequence to a reference genome.}

\item{max.gap}{A maximum length of a gap within a single contig alignments to be collapsed.}

\item{report.mapped.ends}{Set to \code{TRUE} if alignment ends of each contig to the reference should reported as well.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function will take a BED file of contig alignments to a reference genome and converts them
into a \code{\link{GRanges-class}} object. This function also allows to collapse single unique contigs
that are aligned in multiple pieces after the alignment to the reference.
}
\author{
David Porubsky
}
