% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarUtils.R
\name{cigar2ranges}
\alias{cigar2ranges}
\title{Function to load CIGAR string reported in PAF alignments into a set of genomic ranges.}
\usage{
cigar2ranges(
  paf.table = NULL,
  coordinate.space = "reference",
  min.insertion.size = 50,
  min.deletion.size = 50,
  collapse.mismatches = TRUE
)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{coordinate.space}{A used defined coordinate space given CIGAR should be parsed against, either 'reference' or 'query'.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{collapse.mismatches}{Set to \code{TRUE} if mismatches should be collapsed in order expand matched regions.}
}
\value{
A \code{\link{GRanges}} object.
}
\description{
Function to load CIGAR string reported in PAF alignments into a set of genomic ranges.
}
\examples{
## Get PAF to process
paf.file <- system.file("extdata", "test3.paf", package = "SVbyEye")
## Parse CIGAR into a set of genomic ranges
paf.table <- readPaf(paf.file = paf.file)
cigar2ranges(paf.table = paf.table)

}
\author{
David Porubsky
}
