% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequenceContent.R
\name{fasta2nucleotideContent}
\alias{fasta2nucleotideContent}
\title{Get sequence content from a single FASTA file.}
\usage{
fasta2nucleotideContent(
  fasta.seq = NULL,
  fasta.file = NULL,
  binsize = NULL,
  sequence.pattern = NULL,
  nucleotide.content = NULL
)
}
\arguments{
\item{fasta.seq}{A \code{\link{DNAStringSet-class}} object containing a single FASTA sequence.}

\item{fasta.file}{A path to a valid FASTA file to be processed.}

\item{binsize}{A user defined number of base pairs to split the loaded FASTA sequence into}

\item{sequence.pattern}{A user defined DNA sequence pattern which occurrences will be counted in submitted `fasta.seq`.
(e.g. set `sequence.pattern` to 'GA' to obtain counts of all 'GA' occurrences per FASTA sequence).}

\item{nucleotide.content}{A user defined nucleotides which total content will be counted in submitted `fasta.seq`.
(e.g. to obtain 'GC' content set `nucleotide.content` to 'GC')}
}
\value{
A \code{\link{GRanges-class}} object with reported nucleotide content metacolumns
}
\description{
This function with takes a path to FASTA file containing a single sequence and report counts and frequencies of user defined
`sequence.pattern` (such as exact DNA pattern, e.g. 'GA') or `nucleotide.content` (such as sequence GC content) across the
whole sequence or in user defined bins.
}
\examples{
## Get FASTA to process ##
asm.fasta <- system.file("extdata", "test_getFASTA_query.fasta", package = "SVbyEye")
## Report sequence content for a given FASTA file in 5000bp long bins
fasta2nucleotideContent(fasta.file = asm.fasta, binsize = 5000, sequence.pattern = "AT")

}
\author{
David Porubsky
}
