% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPaf.R
\name{filterPaf}
\alias{filterPaf}
\title{Filter PAF alignments}
\usage{
filterPaf(
  paf.table,
  min.mapq = 10,
  min.align.len = 1000,
  min.align.n = 1,
  target.region = NULL,
  query.region = NULL,
  drop.self.align = TRUE
)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{min.mapq}{Minimum mapping quality to retain.}

\item{min.align.len}{Minimum alignment length to retain.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{target.region}{A user defined target region to load in a character string ('chr:start-end') or as
a \code{\link{GRanges-class}} object containing a single genomic region.}

\item{query.region}{A user defined query region to load in a character string ('chr:start-end') or as
a \code{\link{GRanges-class}} object containing a single genomic region.}

\item{drop.self.align}{Remove alignments of a given sequence to itself.}
}
\value{
A \code{tibble} of filtered PAF alignments.
}
\description{
This function takes loaded PAF alignments from \function{readPaf} and perform
user defined filtering of input alignments based on user defined region, mapping quality, and
alignment length (etc.).
}
\author{
David Porubsky
}
