% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipPaf.R
\name{flipPaf}
\alias{flipPaf}
\title{Flip PAF alignments}
\usage{
flipPaf(paf.table, majority.strand = NULL, force = FALSE)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{force}{Set to \code{TRUE} if query PAF alignments should be flipped.}
}
\value{
A \code{tibble} of PAF alignments flipped based on desired majority strand orientation.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and flips 
the alignments given the desired majority strand orientation (Either '+' or '-').
}
\author{
David Porubsky
}
