% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipPaf.R
\name{flipPaf}
\alias{flipPaf}
\title{Flip orientation of PAF alignments.}
\usage{
flipPaf(paf.table, majority.strand = NULL, force = FALSE, flip.seqnames = NULL)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{force}{Set to \code{TRUE} if query PAF alignments should be flipped.}

\item{flip.seqnames}{User defined query or target sequence IDs to be flipped in orientation.}
}
\value{
A \code{tibble} of PAF alignments flipped based on desired majority strand orientation.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and flips
the orientation of PAF alignments given the desired 'majority.strand' orientation (Either '+' or '-').
}
\examples{
## Get PAF to process
paf.file <- system.file("extdata", "test1.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Force to flip PAF alignments
flipPaf(paf.table = paf.table, force = TRUE)

}
\author{
David Porubsky
}
