% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arrowhead.R
\name{geom_arrowhead}
\alias{geom_arrowhead}
\title{A 'ggplot2' geom to draw genomic annotations as arrowheads}
\usage{
geom_arrowhead(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  arrowhead_height = grid::unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2.}

\item{arrowhead_height}{\code{grid::unit()} object giving the height of the
arrowhead.  Defaults to 3 mm.}
}
\description{
\code{geom_arrowhead()} draws genomic ranges as arrowheads, allowing to draw for instance segmental
duplication maps.
}
\details{
This geom draws genes as arrows along a horizontal line representing the
molecule. The start and end locations of the gene are expressed with the
\code{xmin} and \code{xmax} aesthetics, while the molecule can be specified with the
\code{y} aesthetic. Optionally, an additional \code{forward} aesthetic can be used to
reverse the orientation of some or all genes from that implied by \code{xmin} and
\code{xmax}.

Unless the plot is faceted with a free x scale, all the molecules will share
a common x axis. This means that if the locations are very different across
different molecules, the genes might appear very small and squished together
with a lot of unnecessary empty space. To get around this, either facet the
plot with \code{scales = "free_x"}, or normalise the gene locations if their
exact locations are not important.

See \code{make_alignment_dummies()} for a method to align genes between molecules.
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (start and end of the gene; will be used to determine gene
orientation)
\item y (molecule)
\item forward (if any value that is not TRUE, or coercible to TRUE, the gene
arrow will be drawn in the opposite direction to that determined by \code{xmin}
and \code{xmax})
\item alpha
\item colour
\item fill
\item linetype
\item size
}
}

\examples{

ggplot2::ggplot(example_genes, ggplot2::aes(xmin = start, xmax = end,
                                            y = molecule, fill = gene)) +
geom_arrowhead() +
ggplot2::facet_wrap(~ molecule, scales = "free")

}
