% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_miropeats.R
\name{geom_miropeats}
\alias{geom_miropeats}
\title{A 'ggplot2' geom to draw genomic alignments in a miropeats style.}
\usage{
geom_miropeats(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  stat = "miropeats",
  position = "identity",
  n = 100,
  na.rm = FALSE,
  strength = 0.5,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, }{etc... As standard for ggplot2.}
}
\description{
\code{geom_miropeats()} draws miropeat style polygons between query to target genomic alignments.
}
\details{
This geom draws polygons between query to target alignments defined in PAF format.
Such alignments are first loaded using \code{\link{readPaf}} function into a \code{tibble} object.
Then loaded alignments are converted into the plotting coordinates using \code{\link{paf2coords}} function.
molecule. Resulting \code{data.frame} object contains required aesthetics \code{x} and \code{y} coordinates as well as
\code{group} field that is required in order to determine which coordinates represent a single alignment.
}
\section{Aesthetics}{

\code{geom_miropeats()} require or can take the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{group}
\item color
\item linewidth
\item linetype
\item alpha
\item fill
\item size
}
}

\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF 
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Convert PAF alignments to plotting coordinates
coords <- paf2coords(paf.table = paf.table)
## Make a plot 
ggplot2::ggplot(coords) +
 geom_miropeats(ggplot2::aes(x, y, group = group, fill = direction), alpha = 0.5)

}
\seealso{
\code{\link[=readPaf]{readPaf()}}, \code{\link[=paf2coords]{paf2coords()}}
}
\author{
David Porubsky
}
