% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wide_arc.R
\name{geom_wide_arc}
\alias{geom_wide_arc}
\title{A 'ggplot2' geom to draw arcs between genomic alignments.}
\usage{
geom_wide_arc(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  stat = "wide_arc",
  position = "identity",
  n = 100,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As is standard for ggplot2.}

\item{geom}{The geometric object to use to display the data, either as a
\code{ggproto} \code{Geom} subclass or as a string naming the geom stripped of the
\code{geom_} prefix (e.g. \code{"point"} rather than \code{"geom_point"})}

\item{n}{The number of points to create for each alignment polygon (Default : `100`).}
}
\description{
`geom_wide_arc()` draws wide polygons between two sets of start and end coordinates.
}
\details{
This geom is intended to draws wide arc polygons between self-alignments defined in PAF format.
Such alignments can be directly visualized using a wrapper function \code{\link{plotSelf}}.
Input data is a \code{data.frame} object that contains required aesthetics `x` and `y` coordinates as well as
`group` field that is required in order to determine which coordinates represent a single alignment.
}
\section{Aesthetics}{

`geom_wide_arc()` require or can take the following aesthetics
(required aesthetics are in bold):

- **x**
- **y**
- **group**
- color
- linewidth
- linetype
- alpha
- fill
- size
}

\examples{
## Create example data.frame to plot
plt.df <- data.frame(
    x = c(100, 500, 200, 1000),
    y = 0,
    group = 1
)
## Make a plot
ggplot2::ggplot(plt.df) +
    geom_wide_arc(ggplot2::aes(x = x, y = y, group = group), alpha = 0.5)

}
\seealso{
[plotSelf()]
}
\author{
David Porubsky
}
