% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftRangesToAlignment.R
\name{liftRangesToAlignment}
\alias{liftRangesToAlignment}
\title{Function to lift coordinates to the alignment in PAF format.}
\usage{
liftRangesToAlignment(gr = NULL, paf.file = NULL, direction = "query2target")
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object containing single or multiple ranges in query or target sequence coordinates.}

\item{paf.file}{A path to a PAF file containing alignments of a query sequence to a target sequence.}

\item{direction}{One of the possible, lift ranges from query to target 'query2target' or vice versa 'target2query'.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Function to lift coordinates to the alignment in PAF format.
}
\examples{
## Define range(s) to lift
roi.gr <- as('chr17:46645907-46697277', 'GRanges')
## Get PAF alignment to lift to
paf.file <- system.file("extdata", "test_lift1.paf", package="SVbyEye")
## Lift target range to query coordinates
liftRangesToAlignment(gr = roi.gr, paf.file = paf.file, direction = 'target2query')

}
\author{
David Porubsky
}
