% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftRangesToAlignment.R,
%   R/liftRangesToAlignment_deprecated.R
\name{liftRangesToAlignment}
\alias{liftRangesToAlignment}
\title{Function to lift coordinates to the alignment in PAF format.}
\usage{
liftRangesToAlignment(gr = NULL, paf.file = NULL)

liftRangesToAlignment(gr = NULL, paf.file = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object containing single or multiple ranges in local sequence coordinates.}

\item{paf.file}{A path to a PAF file containing alignments of a local sequence to the reference sequence.}

\item{direction}{One of the possible, lift ranges from query to target 'query2target' or vice versa 'target2query'.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.

A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Function to lift coordinates to the alignment in PAF format.

Function to lift local sequence coordinates to the alignment of a given sequence to the reference.
}
\author{
David Porubsky
}
