% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftRangesToAlignment.R
\name{liftRangesToAlignment}
\alias{liftRangesToAlignment}
\title{Function to lift coordinates to the alignment in PAF format.}
\usage{
liftRangesToAlignment(gr = NULL, paf.file = NULL, direction = "query2target")
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object containing single or multiple ranges in query or target sequence coordinates.}

\item{paf.file}{A path to a PAF file containing alignments of a query sequence to a target sequence.}

\item{direction}{One of the possible, lift ranges from query to target 'query2target' or vice versa 'target2query'.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Function to lift coordinates to the alignment in PAF format.
}
\author{
David Porubsky
}
