% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskFASTA.R
\name{maskFasta}
\alias{maskFasta}
\title{Mask FASTA sequence at defined regions.}
\usage{
maskFasta(
  fasta.file,
  mask.ranges = NULL,
  invert = FALSE,
  mask.character = "N",
  fasta.save = NULL
)
}
\arguments{
\item{fasta.file}{A path to a FASTA file to be masked.}

\item{mask.ranges}{A \code{\link{IRanges-class}} object of coordinates to be masked in input FASTA.}

\item{invert}{If set \code{TRUE} ranges defined in 'mask.ranges' will be kept while the rest of the FASTA will be masked.}

\item{mask.character}{A single character to be used for masking \link{default: 'N'}.}

\item{fasta.save}{A path to a filename where to store final FASTA file.}
}
\description{
Mask FASTA sequence at defined regions.
}
\author{
David Porubsky
}
