% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafToCoords.R
\name{paf2coords}
\alias{paf2coords}
\title{Read and filter PAF input file}
\usage{
paf2coords(paf.table)
}
\value{
A \code{data.frame} of PAF alignments reported as x and y coordinate values.
}
\description{
This function takes PAF output file from minimap2 alignments, loads the file and
perform user defined filtering of input alignments based on mapping quality and
alignment length.
}
\author{
David Porubsky
}
