% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{paf2coords}
\alias{paf2coords}
\title{Read and filter PAF input file}
\usage{
paf2coords(
  paf.file,
  min.mapq = 10,
  min.align.len = 1000,
  min.align.n = 1,
  target.region = NULL,
  seqnames.grep = NULL,
  drop.self.align = TRUE
)
}
\arguments{
\item{paf.file}{A path to a PAF file containing alignments to be loaded.}

\item{min.mapq}{Minimum mapping quality to retain.}

\item{min.align.len}{Minimum alignment length to retain.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{target.region}{A user defined target region to load in a character string ('chr:start-end') or as
a \code{\link{GRanges-class}} object containing a single genomic region.}

\item{drop.self.align}{Remove alignments of a given sequence to itself.}

\item{seqname.grep}{Retain only a specific sequence/region name with a given character string.}
}
\value{
A \code{data.frame} of PAF alignments reported as x and y coordinate values.
}
\description{
This function takes PAF output file from minimap2 alignments, loads the file and
perform user defined filtering of input alignments based on mapping quality and
alignment length.
}
\author{
David Porubsky
}
