% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafToCoords.R
\name{paf2coords}
\alias{paf2coords}
\title{Prepare PAF alignments for plotting.}
\usage{
paf2coords(paf.table)
}
\value{
A \code{data.frame} of PAF alignments reported as x and y coordinate values.
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function. Such alignment could be post-processed
using \code{\link{filterPaf}}, \code{\link{breakPaf}} and \code{\link{flipPaf}} functions. Subsequently such alignments are
expanded in a set of x and y coordinates ready to be plotted by \code{\link{geom_miropeats}} function.
}
\author{
David Porubsky
}
