% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafToFasta.R
\name{paf2FASTA}
\alias{paf2FASTA}
\title{Export FASTA sequences from a set of alignments reported in PAF formatted file.}
\usage{
paf2FASTA(
  paf.file,
  bsgenome = NULL,
  asm.fasta = NULL,
  majority.strand = "+",
  revcomp = NULL,
  report.longest.aln = FALSE,
  report.query.name = NULL,
  concatenate.aln = TRUE,
  fasta.save = NULL,
  return = "fasta"
)
}
\arguments{
\item{paf.file}{paf.file A path to a PAF file containing alignments to be loaded.}

\item{bsgenome}{A \code{\link{BSgenome-class}} object of reference genome to get the genomic sequence from.}

\item{asm.fasta}{An assembly FASTA file to extract DNA sequence determined by 'gr' parameter.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{revcomp}{If set to \code{TRUE} FASTA sequence will be reverse complemented.}

\item{report.longest.aln}{If set to \code{TRUE} only the sequence with the most aligned bases will be reported in final FASTA file.}

\item{report.query.name}{A single query (contig) name/id to be reported as FASTA sequence.}

\item{concatenate.aln}{Set to \code{TRUE} if multiple aligned contigs should be concatenated by 100 N's in to a single FASTA sequence \link{default: TRUE}.}

\item{fasta.save}{A path to a filename where to store final FASTA file.}

\item{return}{Set to either 'fasta' or 'index' to return either FASTA in \code{\link{DNAStringSet-class}} object or region index in \code{\link{GRanges-class}} object is returned.}
}
\description{
Export FASTA sequences from a set of alignments reported in PAF formatted file.
}
\author{
David Porubsky
}
