% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafAlignmentToBins.R
\name{pafAlignmentToBins}
\alias{pafAlignmentToBins}
\title{Function to break PAF alignment into matching bases between query and target sequence.
In addition, locations of inserted bases in query and target sequence can be reported as well.}
\usage{
pafAlignmentToBins(paf.aln = NULL, binsize = 10000)
}
\arguments{
\item{paf.aln}{A \code{data.frame} or \code{tibble} containing a single PAF record with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{binsize}{A size of a bin in base pairs to split a PAF alignment into.}
}
\value{
A \code{tibble} object storing binned PAF alignments.
}
\description{
Function to break PAF alignment into matching bases between query and target sequence.
In addition, locations of inserted bases in query and target sequence can be reported as well.
}
\examples{
## Get PAF to bin ##
paf.file <- system.file("extdata", "test3.paf", package = "SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = "cg")
## Split a single PAF alignment into user defined bins
pafAlignmentToBins(paf.aln = paf.table[1, ], binsize = 100)

}
\author{
David Porubsky
}
