% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafAlignmentToBins.R
\name{pafToBins}
\alias{pafToBins}
\title{A wrapper function for \code{\link{pafAlignmentToBins}} binning multiple PAF alignments
into a set of binned alignments between query and target sequence.}
\usage{
pafToBins(paf.table = NULL, binsize = 10000)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{binsize}{A size of a bin in base pairs to split a PAF alignment into.}
}
\value{
A \code{list} of \code{tibble} objects storing matched ('M') alignments as well as structurally variable ('SV') bases if 'report.sv' is TRUE.
}
\description{
A wrapper function for \code{\link{pafAlignmentToBins}} binning multiple PAF alignments
into a set of binned alignments between query and target sequence.
}
\examples{
## Get PAF to bin ##
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Split multiple PAF alignments into user defined bins
pafToBins(paf.table = paf.table, binsize = 10000)

}
\author{
David Porubsky
}
