% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarUtils.R
\name{parseCigarString}
\alias{parseCigarString}
\title{Function to parse CIGAR string into a set interval ranges.}
\usage{
parseCigarString(cigar.str = NULL, coordinate.space = "reference")
}
\arguments{
\item{cigar.str}{A character string containing alignment represented as a CIGAR string.}

\item{coordinate.space}{A used defined coordinate space given CIGAR should be parsed against, either 'reference' or 'query'.}
}
\value{
A \code{\link{IRangesList}} object.
}
\description{
Function to parse CIGAR string into a set interval ranges.
}
\examples{
## Get PAF to process
paf.file <- system.file("extdata", "test3.paf", package="SVbyEye")
## Read in PAF
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Parse CIGAR in reference (target) space
parseCigarString(cigar.str = paf.table$cg)
## Parse CIGAR in query space
parseCigarString(cigar.str = paf.table$cg, coordinate.space = 'query')

}
\author{
David Porubsky
}
