% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{plotAVA}
\alias{plotAVA}
\title{Plot all-versus-all alignments stored in PAf format.}
\usage{
plotAVA(
  paf.file = NULL,
  seqnames.order = NULL,
  seqnames.grep = NULL,
  target.region = NULL,
  min.align.n = 1,
  min.mapq = 0,
  min.align.len = 1000,
  drop.self.align = TRUE,
  majority.strand = NULL,
  outline.miro = FALSE
)
}
\arguments{
\item{paf.file}{A path to a PAF file containing alignments to be loaded.}

\item{seqnames.order}{A user defined order sequence names to be plotted from top to the bottom.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{outline.miro}{Set to \code{TRUE} if separate miropeat segments should be outlined.}
}
\value{
A \code{list} of miropeat style plots.
}
\description{
This function takes PAF output file from minimap2 reporting all-versus-all alignments of multiple FASTA sequences
and visualize the alignments in a miropeat style.
}
\author{
David Porubsky
}
