% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{plotMiro}
\alias{plotMiro}
\title{Plot PAF file as miropeat alignments}
\usage{
plotMiro(
  paf.file = paf.file,
  min.mapq = 10,
  min.align.len = 100,
  min.align.n = 1,
  sd.annot = NULL
)
}
\arguments{
\item{paf.file}{A \code{data.frame} containing x and y coordinates.}

\item{min.mapq}{Minimum mapping quality to retain for PAF alignment file.}

\item{min.align.len}{Minimum alignment length to retain for PAF alignment file.}

\item{min.align.n}{Minimum number of alignments between a pair of sequences/regions.}

\item{sd.annot}{A \code{\link[GenomicRanges]{GRanges}} object containing segmental duplication coordinates.}
}
\value{
A \code{list} of miropeat style plots.
}
\description{
This function takes PAF output file from minimap2 alignemts, and visualize the alignments
in miropeat style.
}
\author{
David Porubsky
}
