% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiro_wrapper.R
\name{plotMiro}
\alias{plotMiro}
\title{Visualize PAF alignments.}
\usage{
plotMiro(
  paf.table,
  min.deletion.size = NULL,
  min.insertion.size = NULL,
  highlight.sv = NULL,
  color.by = "direction"
)
}
\arguments{
\item{paf.table}{A \code{data.frame} or \code{tibble} containing a single or multiple PAF record(s) with 12 mandatory columns
along with CIGAR string defined in 'cg' column.}

\item{min.deletion.size}{A minimum size (in base pairs) of a deletion to be retained.}

\item{min.insertion.size}{A minimum size (in base pairs) of an insertion to be retained.}

\item{highlight.sv}{Visualize alignment embedded structural variation either as an outlined ('outline') or filled ('fill') miropeats.}

\item{color.by}{Color alignments either by directionality ('direction') or fraction of matched base pairs ('fraction.matches').}
}
\value{
A \code{ggplot2} object or \code{list} of  \code{ggplot2} objects.
}
\description{
This function takes PAF output file from minimap2 alignments, and visualize the alignments
in miropeat style.
}
\examples{
## Get PAF to plot
paf.file <- system.file("extdata", "test1.paf", package="SVbyEye")
## Read in PAF 
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
## Optional steps include PAF filtering and flipping query coordinates (see filterPaf and flipPaf function documentation)
## Make a plot
## Color by alignment directionality
plotMiro(paf.table = paf.table, color.by = 'direction')
## Color by fraction of matched bases in each alignment
plotMiro(paf.table = paf.table, color.by = 'fraction.matches')
## Highlight structural variants
paf.file <- system.file("extdata", "test3.paf", package="SVbyEye")
paf.table <- readPaf(paf.file = paf.file, include.paf.tags = TRUE, restrict.paf.tags = 'cg')
plotMiro(paf.table = paf.table, min.deletion.size=50, highlight.sv='outline')

}
\author{
David Porubsky
}
