% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{q2t}
\alias{q2t}
\title{Function to convert between different coordinate scales}
\usage{
q2t(x, q.range, t.range)
}
\arguments{
\item{x}{A \code{vector} of coordinate values to be rescaled.}

\item{q.range}{A \code{vector} containing min and max value for a query range.}

\item{t.range}{A \code{vector} containing min and max value for a target range.}
}
\value{
A \code{vector} of rescaled coordinate values.
}
\description{
This function takes as input query coordinates and a query range and convert them into
coordinate scale defined by a target range.
}
\author{
David Porubsky
}
