% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplots.R
\name{selfdotplot}
\alias{selfdotplot}
\title{Make a horizontal self-alignment dot plot.}
\usage{
selfdotplot(
  aln.coords = NULL,
  format = "nucmer",
  shape = "segment",
  min.align.len = 1000,
  min.align.dist = 1000,
  collapse.overlaps = TRUE,
  highlight.pos = NULL,
  highlight.region = NULL,
  title = NULL,
  return = "plot"
)
}
\arguments{
\item{aln.coords}{A path to a file containing self-alignment coordinates.}

\item{format}{Define format of the input alignment coordinates as either 'nucmer' or minimap2 'mm2', \link{default: 'nucmer'}}

\item{shape}{A shape used to plot aligned sequences: Either 'segment' or 'arc'.}

\item{min.align.dist}{Keep alignment pairs with this or larger distance from each other.}

\item{collapse.overlaps}{Set to \code{TRUE} to merge overlapping pair of alignments with the same relative orientation.}

\item{highlight.pos}{A single or a set of positions to be highlighted as vertical solid lines.}

\item{highlight.region}{A pair of positions to be highlighted as vertical range.}

\item{title}{A title to be added to the final dotplot.}

\item{return}{Define desired output, either alignment plot as 'plot' or self-alignments in stored \code{\link{GRanges-class}} object as 'selfaln', \link{default: 'plot'}}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes self-alignment coordinates from 'nucmer' or 'minimap2' aligner and
construct a horizontal dot plot.
}
\author{
David Porubsky
}
