% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplots.R
\name{selfdotplot}
\alias{selfdotplot}
\title{Make a horizontal self-alignment dot plot.}
\usage{
selfdotplot(
  aln.coords = NULL,
  format = "nucmer",
  min.align.len = 1000,
  min.align.dist = 1000,
  collapse.overlaps = TRUE
)
}
\arguments{
\item{aln.coords}{A path to a file containing self-alignment coordinates.}

\item{format}{Define format of the input alignment coordinates as either 'nucmer' or minimap2 'mm2', \link{default: 'nucmer'}}

\item{min.align.len}{Minimum alignment length to retain.}

\item{min.align.dist}{Keep alignment pairs with this or larger distance from each other.}

\item{collapse.overlaps}{Set to \code{TRUE} to merge overlapping pair of alignments with the same relative orientation.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes self-alignment coordinates from 'nucmer' or 'minimap2' aligner and
construct a horizontal dot plot.
}
\author{
David Porubsky
}
