% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDotp.R
\name{simpledotplot}
\alias{simpledotplot}
\title{Plot simple dotplot of two sequences.}
\usage{
simpledotplot(
  coords.file = NULL,
  format = "nucmer",
  shape = "segment",
  min.align.len = 1000,
  keep.longest.aln = FALSE,
  highlight.pos = NULL,
  highlight.region = NULL,
  genome.coord = FALSE
)
}
\arguments{
\item{coords.file}{A path to a PAF or NUCMER specific file containing alignments to be loaded.}

\item{format}{Is either 'mm2' or 'nucmer' for minimap2 or NUCmer specific input 'coords.file', respectively.}

\item{shape}{A shape used to plot aligned sequences: Either 'segment' or 'point'.}

\item{min.align.len}{Minimum alignment length to retain.}

\item{keep.longest.aln}{Set to \code{TRUE} if a target sequence with a single longest alignment should be kept.}

\item{highlight.pos}{A single or a set of positions to be highlighted as vertical solid lines.}

\item{highlight.region}{A pair of positions to be highlighted as vertical range.}

\item{genome.coord}{Set to \code{TRUE} if the target sequence should be reported in genomic coordinates.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes alignment coordinates from 'nucmer' or 'minimap2' aligner and
construct a simple dot plot.
}
\examples{
## Get NUCmer alignment coordinates to plot
nucmer.coords <- system.file("extdata", "HG00438_1_nucmer.coords", package = "SVbyEye")
## Make a dotplot
simpledotplot(coords.file = nucmer.coords, format = "nucmer")
## Get minimap2 alignments to plot
mm2.coords <- system.file("extdata", "HG01071_2_mm2.paf", package = "SVbyEye")
## Make a dotplot
simpledotplot(coords.file = mm2.coords, format = "mm2")
## Highlight desired genomic positions
highlight.pos <- c(330000, 560000)
simpledotplot(coords.file = mm2.coords, format = "mm2", highlight.pos = highlight.pos)

}
\author{
David Porubsky
}
