% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncRangesDir.R
\name{syncRangesDir}
\alias{syncRangesDir}
\title{Make sure that directionality of set of ranges reflect desired majority strand}
\usage{
syncRangesDir(ranges, majority.strand = "+", strand.only = FALSE)
}
\arguments{
\item{ranges}{A \code{\link{GRanges}} object.}

\item{majority.strand}{A desired majority strand directionality to be reported.}

\item{strand.only}{If set to \code{TRUE} simple character string of strand directionality ('+', '-') is reported.}
}
\description{
Make sure that directionality of set of ranges reflect desired majority strand
}
\author{
David Porubsky
}
